+++ streamlit_webrtc/frontend/src/DeviceSelect/components/DeviceSelectMessage.tsx
@@ -1,0 +1,16 @@
+import Paper from "@mui/material/Paper";
+import { styled } from "@mui/material/styles";
+
+const StyledPaper = styled(Paper)(({ theme }) => ({
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
+  width: "100%",
+  height: "100%",
+  padding: theme.spacing(2),
+  boxSizing: "border-box",
+}));
+
+const DeviceSelectMessage = StyledPaper;
+
+export default DeviceSelectMessage;