+++ streamlit_webrtc/frontend/src/DeviceSelect/components/VideoPreview.tsx
@@ -1,0 +1,10 @@
+import { styled } from "@mui/material/styles";
+
+const StyledVideo = styled("video")({
+  maxWidth: "100%",
+  maxHeight: "100%",
+});
+
+const VideoPreview = StyledVideo;
+
+export default VideoPreview;