+++ streamlit_webrtc/frontend/src/DeviceSelect/components/VoidVideoPreview.stories.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { ComponentStory, ComponentMeta } from "@storybook/react";
+
+import VoidVideoPreview from "./VoidVideoPreview";
+
+export default {
+  title: "DeviceSelect/VoidVideoPreview",
+  component: VoidVideoPreview,
+} as ComponentMeta<typeof VoidVideoPreview>;
+
+export const Default: ComponentStory<typeof VoidVideoPreview> = () => (
+  <VoidVideoPreview />
+);