+++ streamlit_webrtc/frontend/src/DeviceSelect/utils.ts
@@ -1,0 +1,4 @@
+export function stopAllTracks(stream: MediaStream) {
+  stream.getVideoTracks().forEach((track) => track.stop());
+  stream.getAudioTracks().forEach((track) => track.stop());
+}